requirements.txt
frappe
pytesseract
Pillow
requests
openai
google-generativeai
anthropic